% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOIanalyseR.R
\docType{package}
\name{AOIanalyseR}
\alias{AOIanalyseR}
\alias{AOIanalyseR-package}
\title{AOIanalyseR: Analysing eye tracking data with "AOI Sequence Charts"}
\description{
Create "AOI Sequence Charts" and other useful statistics (i.e. aoi view time)
related to your eye tracking data. Compare charts and get information about how
the subjects viewed your graphics, texts, advertisments...
}
\details{
AOI = area of interest. "AOI Sequence Charts" create readable and compareable
timelines out of the fixations of your test subjects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LordSydow/AOIanalyseR}
}

}
\author{
\strong{Maintainer}: Steve Sydow \email{steve.sydow@fh-stralsund.de}

Other contributors:
\itemize{
  \item Paul Wolf \email{paul.wolf@hochschule-stralsund.de} [contributor, thesis advisor]
}

}
