% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_charts.R
\name{createAOISequenceChart}
\alias{createAOISequenceChart}
\title{Create AOI Sequence Charts for all subjects}
\usage{
createAOISequenceChart(
  data,
  scaling = 100,
  size = 7,
  aoi_columns = NULL,
  ggplot_object = NULL
)
}
\arguments{
\item{data}{imported data from \code{importData()}}

\item{scaling}{scale_x_continuous(breaks(), by = scaling)}

\item{size}{width of every bar}

\item{aoi_columns}{order of the aoi}

\item{ggplot_object}{additional ggplot code/settings if needed}
}
\value{
ggplotly chart
}
\description{
Create AOI Sequence Charts for all subjects
}
\examples{
createAOISequenceChart(importedData)
}
\author{
Steve Sydow
}
