% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_charts.R
\name{createAOISplitChart}
\alias{createAOISplitChart}
\title{Create AOI Sequence Chart for one subject}
\usage{
createAOISplitChart(data, name, ggplot_object = NULL)
}
\arguments{
\item{data}{imported data from \code{importData()}}

\item{name}{name of the subject}

\item{ggplot_object}{additional ggplot code/settings if needed}
}
\value{
ggplotly chart
}
\description{
Generates another version of an AOI Sequence Chart for on subject
}
\examples{
createAOISplitChart(data = AOIanalyseR::importedData, name = "subject1")

}
\author{
Steve Sydow
}
