% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{createStats}
\alias{createStats}
\title{Create statistics for the subjects}
\usage{
createStats(data)
}
\arguments{
\item{data}{imported data from \code{importData()}}
}
\value{
\code{Data.Frame}
}
\description{
Shows absolute and relative aoi times
}
\examples{
createStats(importedData)

}
\author{
Steve Sydow
}
