% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{createStatsChart}
\alias{createStatsChart}
\title{Create Bar Chart for aoi view times}
\usage{
createStatsChart(stats, aoi = "all", ggplot_object = NULL)
}
\arguments{
\item{stats}{data from \code{createStats()}}

\item{aoi}{showing all aoi at ones (default) or only one aoi}

\item{ggplot_object}{optional, additional ggplot code/settings}
}
\value{
ggplotly chart
}
\description{
Create Bar Chart for aoi view times
}
\examples{
createStatsChart(createStats(importedData))
createStatsChart(createStats(importedData), aoi = "title")
createStatsChart(createStats(importedData),
                 ggplot_object = ggplot2::labs(title =""))

}
\author{
Steve Sydow
}
