% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importData}
\alias{importData}
\title{Import eye tracking data}
\usage{
importData(
  files = "",
  delimiter = "\\t",
  name_column = "",
  time_column = "",
  aoi_columns = ""
)
}
\arguments{
\item{files}{a Vector containing the file name/names}

\item{delimiter}{sets delimiter (i.e. \code{"\t"} for tab-seperated values)}

\item{name_column}{name of the column containing the subjects name}

\item{time_column}{name of the column containing the times of measurement}

\item{aoi_columns}{names (2 or more) of the columns containing aoi data}
}
\value{
\code{Data.Frame}
}
\description{
Imports and Optimizes data for generating aoi sequence charts
}
\examples{
\dontrun{importData(files = system.file("extdata",
                                        "Sample.tsv",
                                        package = "AOIanalyseR",
                                        mustWork = TRUE),
                    name_column = "subject",
                    time_column = "time",
                    aoi_columns = c("title", "pic", "text"))}

}
\author{
Steve Sydow
}
