% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importedData}
\alias{importedData}
\title{importedData: Imported eye tracking dataset Sample.tsv}
\format{data.frame created with importData-function and the Sample.tsv:
\describe{
  \item{Name}{name of the test subject}
  \item{AOI}{the aoi the subject is fixating}
  \item{Time_Start}{start of the aoi}
  \item{Time_End}{end of the aoi}
  \item{Time_Total}{total view time}
}}
\source{
Randomly generated
}
\usage{
importedData
}
\description{
Contains the necessary data to create statistics and "AOI Sequence Charts"
}
\keyword{datasets}
