% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{oneAOIColumn}
\alias{oneAOIColumn}
\title{Unite aoi columns into one column}
\usage{
oneAOIColumn(data, aoi_columns)
}
\arguments{
\item{data}{imported data from \code{importData()}}

\item{aoi_columns}{names of the columns containing aoi data}
}
\value{
\code{Data.Frame}
}
\description{
Create a column containing all the viewed aoi
}
\author{
Steve Sydow
}
