% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rleAOIData}
\alias{rleAOIData}
\title{Run length encoding (rle)}
\usage{
rleAOIData(data)
}
\arguments{
\item{data}{optimized data from \code{oneAOIColumn()}}
}
\value{
\code{Data.Frame}
}
\description{
rle the aoi data
}
\author{
Steve Sydow
}
